rename init to initialize
